pyproject.toml
[project]
name = "sample-backend-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pypdf2>=3.0.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "openai>=1.53.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "streamlit>=1.39.0",
    "pandas>=2.2.0",
    "requests>=2.32.5",
]
